+++ src/ts/load.ts
@@ -8,7 +8,7 @@ import { IBPMInfo } from './bsmap/types/beatmap/v2/bpmInfo';
 import { IAudio } from './bsmap/types/beatmap/v4/audioData';
 import { IBPMEvent } from './bsmap/types/beatmap/v3/bpmEvent';
 import settings from './settings';
-import { Beatmap, loadDifficulty, loadInfo, loadLightshow, toV4Beatmap } from './bsmap/beatmap/mod';
+import { Beatmap, loadDifficulty, loadInfo, loadLightshow } from './bsmap/beatmap/mod';
 import { IWrapBeatmap } from './bsmap/types/beatmap/wrapper/beatmap';
 import { INoteContainer, NoteContainerType } from './types/checks/container';